Support build_web_compilers
