+++ melos.yaml
@@ -51,3 +51,11 @@ scripts:
     # as they could change the behaviour of how tests filter packages.
     env:
       MELOS_TEST: true
+  
+  test_build:
+    description: Runs tests with build_test
+    run: dart run build_runner test -- -p chrome
+    exec:
+      concurrency: 1
+    packageFilters:
+      dependsOn: build_test
