+++ packages/sqlite_async/pubspec.yaml
@@ -1,6 +1,6 @@
 name: sqlite_async
 description: High-performance asynchronous interface for SQLite on Dart and Flutter.
-version: 0.11.2
+version: 0.11.3
 repository: https://github.com/powersync-ja/sqlite_async.dart
 environment:
   sdk: ">=3.5.0 <4.0.0"
@@ -21,6 +21,9 @@ dependencies:
   web: ^1.0.0
 
 dev_dependencies:
+  build_runner: ^2.4.14
+  build_web_compilers: ^4.1.1
+  build_test: ^2.2.3
   lints: ^5.0.0
   test: ^1.21.0
   test_api: ^0.7.0
